Implement schema argument flattening and decoding logic
- Added `flatten_schema_arg` function to handle custom types and lists in command arguments.
- Introduced `decode` function for decoding byte arrays into structured data based on custom types.
- Created error handling for various decoding and serialization issues.
- Removed the `executor.rs` file as part of code refactoring.
- Added a new `parser.rs` file to handle command line argument parsing with custom types.
- Updated `test_utils.rs` to include new mock command arguments and test cases.
- Introduced new error types and decoding utilities in `error.rs` and `decoder.rs`.
- Enhanced type handling in `types/mod.rs` for better serialization and deserialization of custom types.
