--- odra-cli/src/executor.rs
@@ -1,90 +1,0 @@
-// use std::{collections::HashMap, str::FromStr};
-
-// use clap::ArgMatches;
-// use odra::{
-//     casper_types::{bytesrepr::Bytes, U512},
-//     contract_def::{Entrypoint, HasIdent},
-//     entry_point_callback::EntryPointsCaller,
-//     host::{EntryPointsCallerProvider, HostEnv},
-//     prelude::Address,
-//     OdraContract
-// };
-
-// use crate::{
-//     args,
-//     cmd::args::{read_arg, ARG_GAS, ARG_PRINT_EVENTS},
-//     entry_point::CallError
-// };
-
-// pub trait OnChainExecutor {
-//     fn execute_contract(
-//         &self,
-//         contract_name: &str,
-//         entry_point: &Entrypoint,
-//         args: &ArgMatches,
-//         contract_address: Address
-//     ) -> Result<Bytes, CallError>;
-// }
-
-// impl OnChainExecutor for HostEnv {
-//     fn execute_contract(
-//         &self,
-//         contract_name: &str,
-//         entry_point: &Entrypoint,
-//         args: &ArgMatches,
-//         contract_address: Address
-//     ) -> Result<Bytes, CallError> {
-//         let amount = read_arg(args, ARG_PRINT_EVENTS, U512::from_dec_str)?;
-
-//         let runtime_args = args::compose(entry_point, args, types)?;
-//         let method = &entry_point.name;
-//         let is_mut = entry_point.is_mutable;
-//         let ty = &entry_point.return_ty;
-//         let call_def = CallDef::new(method, is_mut, runtime_args).with_amount(amount);
-//         let use_proxy = ty.0 != NamedCLType::Unit || !call_def.amount().is_zero();
-
-//         if is_mut {
-//             let gas = read_arg(args, ARG_GAS, FromStr::from_str)?;
-//             self.set_gas(gas);
-//         }
-
-//         let print_events = if is_mut {
-//             args.get_flag(ARG_PRINT_EVENTS)
-//         } else {
-//             false
-//         };
-//         if print_events {
-//             prettycli::info("Syncing events for the call...");
-//         }
-//         self.set_captures_events(print_events);
-//         self.raw_call_contract(contract_address, call_def, use_proxy)
-//             .map_err(|e| CallError::ExecutionError {
-//                 contract_name: contract_name.to_string(),
-//                 method: method.to_string(),
-//                 message: match e {
-//                     OdraError::VmError(VmError::Other(msg)) => msg,
-//                     _ => format!("{:?}", e)
-//                 }
-//             })?;
-//     }
-//     // Implement methods for executing on-chain operations
-// }
-
-// pub struct MockExecutor;
-
-// impl OnChainExecutor for MockExecutor {
-//     fn execute_contract(
-//         &self,
-//         contract_name: &str,
-//         entry_point: &Entrypoint,
-//         args: &ArgMatches,
-//         contract_address: Address
-//     ) -> Result<Bytes, CallError> {
-//         // Mock implementation for testing purposes
-//         let result = Bytes::from(format!(
-//             "Mocked execution of {}::{} with args {:?} at address {:?}",
-//             contract_name, entry_point.name, args, contract_address
-//         ));
-//         Ok(result)
-//     }
-// }
