+++ odra-cli/src/lib.rs
@@ -21,17 +21,15 @@ use odra::{
     OdraContract
 };
 
-mod args;
 mod cmd;
 mod container;
 mod custom_types;
 mod entry_point;
-mod executor;
+mod parser;
 #[cfg(test)]
 mod test_utils;
 mod types;
 
-pub use args::CommandArg;
 pub use cmd::scenario::{ScenarioArgs, ScenarioError};
 pub use container::DeployedContractsContainer;
 use scenario::{Scenario, ScenarioMetadata};
@@ -247,181 +245,3 @@ impl OdraCli {
         cmd.run(&self.host_env, &args, &self.custom_types, contracts_path)
     }
 }
-
-// #[cfg(test)]
-// mod tests {
-//     use std::{cell::RefCell, rc::Rc};
-
-//     use odra::host::HostContext;
-
-//     use super::*;
-
-//     pub struct MockHostCtx;
-
-//     impl HostContext for MockHostCtx {
-//         #[doc = " Sets the caller address for the current contract execution."]
-//         fn set_caller(&self, caller: Address) {
-//             todo!()
-//         }
-
-//         #[doc = " Sets the gas limit for the current contract execution."]
-//         fn set_gas(&self, gas: u64) {
-//             todo!()
-//         }
-
-//         #[doc = " Returns the caller address for the current contract execution."]
-//         fn caller(&self) -> Address {
-//             todo!()
-//         }
-
-//         #[doc = " Returns the account address at the specified index."]
-//         fn get_account(&self, index: usize) -> Address {
-//             todo!()
-//         }
-
-//         #[doc = " Returns the validator public key."]
-//         fn get_validator(&self, index: usize) -> PublicKey {
-//             todo!()
-//         }
-
-//         #[doc = " The validator at the given index will withdraw all funds and be removed from the validator set."]
-//         fn remove_validator(&self, index: usize) {
-//             todo!()
-//         }
-
-//         #[doc = " Returns the CSPR balance of the specified address."]
-//         fn balance_of(&self, address: &Address) -> U512 {
-//             todo!()
-//         }
-
-//         #[doc = " Advances the block time by the specified time difference."]
-//         fn advance_block_time(&self, time_diff: u64) {
-//             todo!()
-//         }
-
-//         #[doc = " Advances the block time by the specified time difference and processes auctions."]
-//         fn advance_with_auctions(&self, time_diff: u64) {
-//             todo!()
-//         }
-
-//         #[doc = " Time between auctions in milliseconds."]
-//         fn auction_delay(&self) -> u64 {
-//             todo!()
-//         }
-
-//         #[doc = " Time for the funds to be transferred back to the delegator after undelegation in milliseconds."]
-//         fn unbonding_delay(&self) -> u64 {
-//             todo!()
-//         }
-
-//         #[doc = " Returns the delegated amount for the specified delegator and validator."]
-//         fn delegated_amount(&self, delegator: Address, validator: PublicKey) -> U512 {
-//             todo!()
-//         }
-
-//         #[doc = " Returns the current block time."]
-//         fn block_time(&self) -> u64 {
-//             todo!()
-//         }
-
-//         #[doc = " Returns the event bytes for the specified contract address and index."]
-//         fn get_event(&self, contract_address: &Address, index: u32) -> Result<Bytes, EventError> {
-//             todo!()
-//         }
-
-//         #[doc = " Returns the native event bytes for the specified contract address and index."]
-//         fn get_native_event(
-//             &self,
-//             contract_address: &Address,
-//             index: u32
-//         ) -> Result<Bytes, EventError> {
-//             todo!()
-//         }
-
-//         #[doc = " Returns the number of emitted events for the specified contract address."]
-//         fn get_events_count(&self, contract_address: &Address) -> Result<u32, EventError> {
-//             todo!()
-//         }
-
-//         #[doc = " Returns the number of emitted native events for the specified contract address."]
-//         fn get_native_events_count(&self, contract_address: &Address) -> Result<u32, EventError> {
-//             todo!()
-//         }
-
-//         #[doc = " Calls a contract at the specified address with the given call definition."]
-//         fn call_contract(
-//             &self,
-//             address: &Address,
-//             call_def: CallDef,
-//             use_proxy: bool
-//         ) -> OdraResult<Bytes> {
-//             todo!()
-//         }
-
-//         #[doc = " Creates a new contract with the specified name, initialization arguments, and entry points caller."]
-//         fn new_contract(
-//             &self,
-//             name: &str,
-//             init_args: RuntimeArgs,
-//             entry_points_caller: EntryPointsCaller
-//         ) -> OdraResult<Address> {
-//             todo!()
-//         }
-
-//         #[doc = " Registers an existing contract with the specified address, name, and entry points caller."]
-//         fn register_contract(
-//             &self,
-//             address: Address,
-//             contract_name: String,
-//             entry_points_caller: EntryPointsCaller
-//         ) {
-//             todo!()
-//         }
-
-//         #[doc = " Returns the contract environment."]
-//         fn contract_env(&self) -> ContractEnv {
-//             todo!()
-//         }
-
-//         #[doc = " Returns the gas report for the current contract execution."]
-//         fn gas_report(&self) -> GasReport {
-//             todo!()
-//         }
-
-//         #[doc = " Returns the gas cost of the last contract call."]
-//         fn last_call_gas_cost(&self) -> u64 {
-//             todo!()
-//         }
-
-//         #[doc = " Signs the specified message with the given address and returns the signature."]
-//         fn sign_message(&self, message: &Bytes, address: &Address) -> Bytes {
-//             todo!()
-//         }
-
-//         #[doc = " Returns the public key associated with the specified address."]
-//         fn public_key(&self, address: &Address) -> PublicKey {
-//             todo!()
-//         }
-
-//         #[doc = " Transfers the specified amount of CSPR from the current caller to the specified address."]
-//         fn transfer(&self, to: Address, amount: U512) -> OdraResult<()> {
-//             todo!()
-//         }
-//     }
-
-//     pub fn env() -> HostEnv {
-//         HostEnv::new(Rc::new(RefCell::new(MockHostCtx)))
-//     }
-
-//     #[test]
-//     fn test_odra_cli_new() {
-//         let cli = OdraCli::test(env());
-//         assert!(cli.main_cmd.main_cmd.get_subcommands().count() == 0);
-//     }
-
-//     #[test]
-//     fn test_odra_cli_build() {
-//         let cli = OdraCli::test(env());
-//         assert!(cli.main_cmd.main_cmd.get_subcommands().count() > 0);
-//     }
-// }
